more small cleanups
